Fix 1.8 forge support
