+++ Forge/resource/META-INF/mods.toml
@@ -1,9 +1,5 @@
 modLoader="javafml"
 loaderVersion="[25,)"
-issueTrackerURL="https://github.com/xfl03/MCCustomSkinLoader/issues"
-displayURL="https://github.com/xfl03/MCCustomSkinLoader"
-credits="GPLv3"
-authors=["xfl03", "JLChnToZ"]
 [[mods]]
 modId="customskinloader"
 version="${modFullVersion}"
@@ -11,6 +7,9 @@ displayName="CustomSkinLoader"
 description='''
 Custom Skin Loader for Minecraft
 '''
+issueTrackerURL="https://github.com/xfl03/MCCustomSkinLoader/issues"
+displayURL="https://github.com/xfl03/MCCustomSkinLoader"
+authors=["xfl03", "JLChnToZ"]
 [[dependencies.customskinloader]]
 modId="forge"
 mandatory=true
