+++ build.properties
@@ -1,6 +1,6 @@
 name=CustomSkinLoader
 group=customskinloader
 version=14.10
-dev_version=
+dev_version=a
 #RELEASE: snapshot=0
-snapshot=0
+snapshot=1
